djangoProject/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.7
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
jwt==1.2.0
Markdown==3.3.4
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
